Changed path variables in TagController to be consistent between mapping and method parameter

+++ src/main/java/com/revature/controller/TagController.java
@@ -7,7 +7,7 @@ import com.revature.model.Tag;
 public interface TagController {
 
 	List<Tag> getAllTags();
-	Tag getTag(int id);
-	void deleteTag(int id);
+	Tag getTag(long id);
+	void deleteTag(long id);
 	long saveTag(Tag tag);
 }
